Add IPC for basic process communication
Added a really bascic  IPC communication
between IO subsytem and CO2 cli commands

A UNIX DOMAIN SOCKET is now provided for
the above comunication.

A simple server in io.py has been added
It will need to be converted in a  UNIX
daemon to be started/stopped in systemd
It'll be responsible to accept messages
from CO2 cli commands and apply them to
the underlying IO system.

 Client			 Socket		 Server
+---------------+	+--------+	+---------------+
| CO2 Commands	|	| UNIX	 |	| 		|
|		|	| 	 |	| 		|
| touch		|	|	 |	| 		|
| rm		|======>| DOMAIN |=====>| IO/Filesystem |
| mkdir		|	|	 |	|		|
| rmdir		|	|	 |	|		|
| ...		|	| SOCKET |	|		|
+---------------+	+--------+	+---------------+